+++ catalog/db_scanner.go
@@ -13,6 +13,7 @@ type DBScanner interface {
 	Next() bool
 	Value() *DBScannerEntry
 	Err() error
+	getBranchID() int64
 }
 
 func ScanDBEntryUntil(s DBScanner, ent *DBScannerEntry, p string) (*DBScannerEntry, error) {
